requirements.txt
neo4j==5.6.0
neo4j_driver==5.5.0
numpy==1.21.5
pandas==1.4.4
Pillow==9.4.0
plotly==5.9.0
plotly_express==0.4.1
streamlit==1.19.0
~treamlit==1.18.1
